fix imbalance calculation

+++ iim_dashboard/islanding/models.py
@@ -13,12 +13,12 @@ class IslandingScheme(models.Model):
 
     method_name                = models.CharField('Method', max_length=20)
     
-    island_imbalance           = ArrayField(models.IntegerField(blank=True, null=True), size=3, null=True)
+    island_imbalance           = ArrayField(models.FloatField(blank=True, null=True), size=3, null=True)
     total_imbalance            = models.FloatField(blank=True, null=True)
     # island_imbalance           = ArrayField(models.FloatField(blank=True, null=True), size=3, null=True)
     # total_imbalance            = models.FloatField(blank=True, null=True)
     
-    island_imbalance_after_cut = ArrayField(models.IntegerField(blank=True, null=True), size=3, null=True)
+    island_imbalance_after_cut = ArrayField(models.FloatField(blank=True, null=True), size=3, null=True)
     total_imbalance_after_cut  = models.FloatField(blank=True, null=True)
     
     lines_to_cut               = models.IntegerField(blank=True, null=True)
